Refactor: Button component
