+++ packages/mikado_reborn/src/components/Button/Button.scss
@@ -1,4 +1,4 @@
-@import '../../assets/styles/styles.scss';
+@import '../../assets/styles/styles';
 
 .mkr__button {
   $button: &;
