+++ packages/mikado_reborn/src/components/Button/Contained/ContainedButton.vue
@@ -11,7 +11,7 @@
 
 <script lang="ts" setup>
 import { withDefaults, defineProps } from 'vue';
-import MkrButton from '../../../mixins/Button/Button.vue';
+import MkrButton from '../Button.vue';
 import './ContainedButton.scss';
 
 type containedButtonThemes = 'primary' | 'primary-light' | 'secondary' | 'secondary-light' | 'danger-light' | 'neutral';
