+++ packages/mikado_reborn/src/components/Button/Outlined/OutlinedButton.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script lang="ts" setup >
-import MkrButton from '../../../mixins/Button/Button.vue';
+import MkrButton from '../Button.vue';
 import './OutlinedButton.scss';
 
 </script>
