+++ packages/mikado_reborn/src/components/Button/Text/TextButton.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script lang="ts" setup>
-import MkrButton from '../../../mixins/Button/Button.vue';
+import MkrButton from '../Button.vue';
 import './TextButton.scss';
 
 import { withDefaults, defineProps } from 'vue';
